requirements.txt
streamlit==1.40.1
fastai==2.7.12
Pillow==10.0.1